move submissions endpoints to their own file
