+++ src/data-layer/main.go
@@ -41,6 +41,7 @@ func main() {
 
 	addUserRoutes(mux)
 	addProblemRoutes(mux)
+	addSubmissionRoutes(mux)
 
 	addr := ":" + *port
 	err = http.ListenAndServe(addr, mux)
