+++ src/data-layer/problems.go
@@ -17,8 +17,6 @@ import (
 func addProblemRoutes(mux *goji.Mux) {
 	mux.HandleFunc(pat.Post("/v1/problems"), postProblem)
 	mux.HandleFunc(pat.Get("/v1/problems/:problem_id"), getProblem)
-	mux.HandleFunc(pat.Post("/v1/problems/:problem_id/submissions"), postSubmission)
-	mux.HandleFunc(pat.Get("/v1/problems/:problem_id/submissions/:submission_id"), getSubmission)
 }
 
 // TODO: get user id from jwt, check for existing user
@@ -37,6 +35,20 @@ func postProblem(w http.ResponseWriter, r *http.Request) {
 		logrus.WithError(err).Error("JSON parse error")
 		w.WriteHeader(http.StatusInternalServerError)
 		fmt.Fprint(w, `{"code":"500", "message":"JSON parse error"}`)
+		return
+	}
+
+	user, err := db.GetUserByID(reqData.UserID)
+	if err != nil {
+		logrus.WithError(err).Error("error checking for existing user")
+		w.WriteHeader(http.StatusInternalServerError)
+		fmt.Fprint(w, `{"code":"500", "message":"error checking for existing user"}`)
+		return
+	}
+	if user == nil {
+		logrus.Warn("user does not exist")
+		w.WriteHeader(http.StatusBadRequest)
+		fmt.Fprint(w, `{"code":"400", "message":"user does not exist"}`)
 		return
 	}
 
@@ -123,156 +135,6 @@ func getProblem(w http.ResponseWriter, r *http.Request) {
 	}
 
 	respJSON, err := json.Marshal(problem)
-	if err != nil {
-		logrus.WithError(err).Error("JSON parse error")
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprint(w, `{"code":"500", "message":"JSON parse error"}`)
-		return
-	}
-	fmt.Fprintf(w, string(respJSON))
-	w.WriteHeader(http.StatusOK)
-}
-
-// TODO: verify that the failed test case IS for the specific problem
-func postSubmission(w http.ResponseWriter, r *http.Request) {
-	problemIdParam := pat.Param(r, "problem_id")
-	problemId, err := strconv.Atoi(problemIdParam)
-	if err != nil {
-		logrus.WithError(err).Error("problem id must be int")
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprint(w, `{"code":"500", "message":"problem id must be int"}`)
-		return
-	}
-
-	reqData := new(Submission)
-	reqBodyBytes, err := io.ReadAll(r.Body)
-	if err != nil {
-		logrus.WithError(err).Error("error reading request body")
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprint(w, `{"code":"500", "message":"error reading request body"}`)
-		return
-	}
-
-	err = json.Unmarshal(reqBodyBytes, reqData)
-	if err != nil {
-		logrus.WithError(err).Error("JSON parse error")
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprint(w, `{"code":"500", "message":"JSON parse error"}`)
-		return
-	}
-
-	if reqData.ProblemID != problemId {
-		logrus.Warn("problem id on request body and path do not match")
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprint(w, `{"code":"400", "message":"problem id on request body and path do not match"}`)
-		return
-	}
-
-	problem, err := db.GetProblemByID(problemId)
-	if err != nil {
-		logrus.WithError(err).Error("error checking for existing problem")
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprint(w, `{"code":"500", "message":"error checking for existing problem"}`)
-		return
-	}
-	if problem == nil {
-		logrus.Warn("problem does not exist")
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprint(w, `{"code":"404", "message":"problem does not exist"}`)
-		return
-	}
-
-	user, err := db.GetUserByID(reqData.UserID)
-	if err != nil {
-		logrus.WithError(err).Error("error checking for existing user")
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprint(w, `{"code":"500", "message":"error checking for existing user"}`)
-		return
-	}
-	if user == nil {
-		logrus.Warn("user does not exist")
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprint(w, `{"code":"404", "message":"user does not exist"}`)
-		return
-	}
-
-	response, err := db.CreateSubmission(reqData)
-	if err != nil {
-		logrus.WithError(err).Error("error inserting problem into db")
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprint(w, `{"code":"500", "message":"error inserting problem into db"}`)
-		return
-	}
-	reqData.ID = response.ID
-	reqData.SubmitTime = response.SubmitTime
-
-	respJSON, err := json.Marshal(reqData)
-	if err != nil {
-		logrus.WithError(err).Error("JSON parse error")
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprint(w, `{"code":"500", "message":"JSON parse error"}`)
-		return
-	}
-	fmt.Fprintf(w, string(respJSON))
-	w.WriteHeader(http.StatusOK)
-}
-
-func getSubmission(w http.ResponseWriter, r *http.Request) {
-	problemIdParam := pat.Param(r, "problem_id")
-	submissionIdParam := pat.Param(r, "submission_id")
-
-	problemId, err := strconv.Atoi(problemIdParam)
-	if err != nil {
-		logrus.WithError(err).Error("problem id must be int")
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprint(w, `{"code":"500", "message":"problem id must be int"}`)
-		return
-	}
-
-	submissionId, err := strconv.Atoi(submissionIdParam)
-	if err != nil {
-		logrus.WithError(err).Error("submission id must be int")
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprint(w, `{"code":"500", "message":"submission id must be int"}`)
-		return
-	}
-
-	problem, err := db.GetProblemByID(problemId)
-	if err != nil {
-		logrus.WithError(err).Error("error checking db for problem")
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprint(w, `{"code":"500", "message":"error checking db for problem"}`)
-		return
-	}
-	if problem == nil {
-		logrus.Warn("problem does not exist")
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprint(w, `{"code":"404", "message":"problem does not exist"}`)
-		return
-	}
-
-	submission, err := db.GetSubmission(submissionId)
-	if err != nil {
-		logrus.WithError(err).Error("error retrieving problem")
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprint(w, `{"code":"500", "message":"error retrieving problem"}`)
-		return
-	}
-	if submission == nil {
-		logrus.Warn("submission not found")
-		w.WriteHeader(http.StatusNotFound)
-		fmt.Fprint(w, `{"code":"404", "message":"problem not found"}`)
-		return
-	}
-
-	if problemId != submission.ProblemID {
-		logrus.Warn("problem id does not match the problem for the submission")
-		w.WriteHeader(http.StatusBadRequest)
-		fmt.Fprint(w, `{"code":"404", "message":"problem id does not match the problem for the submission"}`)
-		return
-	}
-
-	respJSON, err := json.Marshal(submission)
 	if err != nil {
 		logrus.WithError(err).Error("JSON parse error")
 		w.WriteHeader(http.StatusInternalServerError)
