+++ src/data-layer/submissions.go
@@ -1,0 +1,122 @@
+package main
+
+import (
+	"encoding/json"
+	"fmt"
+	"io"
+	"net/http"
+	"strconv"
+
+	"github.com/sirupsen/logrus"
+	"goji.io"
+	"goji.io/pat"
+)
+
+func addSubmissionRoutes(mux *goji.Mux) {
+	mux.HandleFunc(pat.Post("/v1/submissions"), postSubmission)
+	mux.HandleFunc(pat.Get("/v1/submissions/:submission_id"), getSubmission)
+}
+
+// TODO: verify that the failed test case IS for the specific problem
+func postSubmission(w http.ResponseWriter, r *http.Request) {
+	reqData := new(Submission)
+	reqBodyBytes, err := io.ReadAll(r.Body)
+	if err != nil {
+		logrus.WithError(err).Error("error reading request body")
+		w.WriteHeader(http.StatusInternalServerError)
+		fmt.Fprint(w, `{"code":"500", "message":"error reading request body"}`)
+		return
+	}
+
+	err = json.Unmarshal(reqBodyBytes, reqData)
+	if err != nil {
+		logrus.WithError(err).Error("JSON parse error")
+		w.WriteHeader(http.StatusInternalServerError)
+		fmt.Fprint(w, `{"code":"500", "message":"JSON parse error"}`)
+		return
+	}
+
+	problem, err := db.GetProblemByID(reqData.ProblemID)
+	if err != nil {
+		logrus.WithError(err).Error("error checking for existing problem")
+		w.WriteHeader(http.StatusInternalServerError)
+		fmt.Fprint(w, `{"code":"500", "message":"error checking for existing problem"}`)
+		return
+	}
+	if problem == nil {
+		logrus.Warn("problem does not exist")
+		w.WriteHeader(http.StatusBadRequest)
+		fmt.Fprint(w, `{"code":"404", "message":"problem does not exist"}`)
+		return
+	}
+
+	user, err := db.GetUserByID(reqData.UserID)
+	if err != nil {
+		logrus.WithError(err).Error("error checking for existing user")
+		w.WriteHeader(http.StatusInternalServerError)
+		fmt.Fprint(w, `{"code":"500", "message":"error checking for existing user"}`)
+		return
+	}
+	if user == nil {
+		logrus.Warn("user does not exist")
+		w.WriteHeader(http.StatusBadRequest)
+		fmt.Fprint(w, `{"code":"404", "message":"user does not exist"}`)
+		return
+	}
+
+	response, err := db.CreateSubmission(reqData)
+	if err != nil {
+		logrus.WithError(err).Error("error inserting problem into db")
+		w.WriteHeader(http.StatusInternalServerError)
+		fmt.Fprint(w, `{"code":"500", "message":"error inserting problem into db"}`)
+		return
+	}
+	reqData.ID = response.ID
+	reqData.SubmitTime = response.SubmitTime
+
+	respJSON, err := json.Marshal(reqData)
+	if err != nil {
+		logrus.WithError(err).Error("JSON parse error")
+		w.WriteHeader(http.StatusInternalServerError)
+		fmt.Fprint(w, `{"code":"500", "message":"JSON parse error"}`)
+		return
+	}
+	fmt.Fprintf(w, string(respJSON))
+	w.WriteHeader(http.StatusOK)
+}
+
+func getSubmission(w http.ResponseWriter, r *http.Request) {
+	submissionIdParam := pat.Param(r, "submission_id")
+
+	submissionId, err := strconv.Atoi(submissionIdParam)
+	if err != nil {
+		logrus.WithError(err).Error("submission id must be int")
+		w.WriteHeader(http.StatusBadRequest)
+		fmt.Fprint(w, `{"code":"500", "message":"submission id must be int"}`)
+		return
+	}
+
+	submission, err := db.GetSubmission(submissionId)
+	if err != nil {
+		logrus.WithError(err).Error("error retrieving problem")
+		w.WriteHeader(http.StatusInternalServerError)
+		fmt.Fprint(w, `{"code":"500", "message":"error retrieving problem"}`)
+		return
+	}
+	if submission == nil {
+		logrus.Warn("submission not found")
+		w.WriteHeader(http.StatusNotFound)
+		fmt.Fprint(w, `{"code":"404", "message":"problem not found"}`)
+		return
+	}
+
+	respJSON, err := json.Marshal(submission)
+	if err != nil {
+		logrus.WithError(err).Error("JSON parse error")
+		w.WriteHeader(http.StatusInternalServerError)
+		fmt.Fprint(w, `{"code":"500", "message":"JSON parse error"}`)
+		return
+	}
+	fmt.Fprintf(w, string(respJSON))
+	w.WriteHeader(http.StatusOK)
+}
